python3.http.client illegal header check breaks when default of `[{}]` supplied. It needs to be a string or bytes like object.
end of the stack trace is:
```
    if _is_illegal_header_value(values[i]):
TypeError: expected string or bytes-like object
```
This  occurs when document.undefined class is used for the partitionKey
header, which hits this code to set a sensible default. Old code set an
empty list dict `[{}]`, but that breaks in python3. This fixes it for
python3.